mcp-tools/photo-migration/pyproject.toml
[project]
name = "photo-migration"
version = "0.1.0"
description = "MCP server for iOS to Android photo migration"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp>=0.1.0",
    "playwright>=1.40.0",
    "python-dotenv>=1.0.0",
    "Pillow>=10.0.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
photo-migration = "photo_migration.server:main"